requirements.txt
flask
flask-cors
paho-mqtt
psycopg2-binary
python-dotenv
requests
